[label][refactor] has refactor complete enough a label-ized test case is working in local dev -- not yet tested in kube
