+++ ratchet-child/ratchet-child.go
@@ -199,81 +199,105 @@ func getEtcdMetaData(containerid string, setalive bool) (map[string]string) {
 
 }
 
+func associateIDEtcd (containerid string,podname string) error {
+
+  // set "/foo" key with "bar" value
+  // log.Print("Setting '/foo' key with 'bar' value")
+  _, err := kapi.Set(context.Background(), "/ratchet/association/" + podname, containerid, nil)
+  if err != nil {
+    log.Fatal(err)
+    return err
+  } else {
+    // print common key info
+    // log.Printf("Set is done. Metadata is %q\n", resp)
+  }
+
+  return nil
+
+}
+
+func isPairContainerAlive(podname string) string {
+  
+  target_key := "/ratchet/association/" + podname
+  resp_containerid, err := kapi.Get(context.Background(), target_key, nil)
+  if err != nil {
+
+      // ErrorCodeKeyNotFound = Key not found, that's exactly the one we know is good.
+      // So let's log when it's not that.
+      // Passing along on this.
+      /*
+      if (err != client.ErrorCodeKeyNotFound) {
+        logger.Println(fmt.Errorf("isPairContainerAlive - possible missing value %s: %v", target_key, err))
+      }
+      */
+
+    } else {
+      // no error? must be there.
+      return resp_containerid.Node.Value;
+    }
+
+  return ""
+
+}
+
 func ratchet(argif string,containerid string, linki LinkInfo) error {
 
   os.Stderr.WriteString("!trace alive The containerid: " + containerid + "\n")
 
-  var result error
-
-  // Go into a loop determining if we're alive 
-  tries := 0
-  i_am_alive := false
-  for amIAlive(containerid) == false {
-    
-    tries++
-
-    if (DEBUG) {
-      logger(fmt.Sprintf("Am I alive? containerid: %v (%v retries)\n",containerid,tries))
-    }
-
-    // We either timeout, or, we're alive.
-    if (tries >= ALIVE_WAIT_RETRIES || i_am_alive) {
-      return fmt.Errorf("Timeout: could not find this container is alive via metadata in %v tries", tries)
-    }
-
-    // Wait for however long.
-    time.Sleep(ALIVE_WAIT_SECONDS * time.Second)
-  }
+  // We no longer care if we're alive anymore. 
+  // If this is up, we can assume the infra container is good to go.
+  // So all we need to do is associate our containerid with our name.
+  associateIDEtcd(containerid,linki.Pod_name)
 
   // If it's determined that we're alive, now we can see if we're primary.
   // If we're not primary, we can just exit right now.
   // Cause the primary side will add to this pair.
 
-  // Go and pick up metadata about me from etcd.
-  my_meta := getEtcdMetaData(containerid,true)
-
-  if (my_meta["primary"] != "true") {
+  if (linki.Primary != "true") {
     // Ok, we're not primary. So... time to exit.
     if (DEBUG) {
-      logger(fmt.Sprintf("Normal termination, this container is not primary (name: %v, containerid: %v, primary: %v)",my_meta["pod_name"],containerid,my_meta["primary"]))
+      logger(fmt.Sprintf("Normal termination, this container is not primary (name: %v, containerid: %v, primary: %v)",linki.Pod_name,containerid,linki.Primary))
     }
 
     return nil
   }
 
   // Check to see there's a valid pair name.
-  if (len(my_meta["pair_name"]) <= 1) {
+  if (len(linki.Pair_name) <= 1) {
     // That's not good.
-    return fmt.Errorf("Pair name appears to be invalid: %v", my_meta["pair_name"])
+    return fmt.Errorf("Pair name appears to be invalid: %v", linki.Pair_name)
   }
 
   // Now we want to check and see if the pair container is alive.
   // So it's time to go into a loop and do that.
   // if the pair container is alive -- bada bing, we can execute koko.
 
-  for isContainerAlive(my_meta["pair_name"]) == false {
+  var pair_containerid string
+  tries := 0
+
+  for {
     
+    pair_containerid = isPairContainerAlive(linki.Pair_name)
+
+    if (len(pair_containerid) >= 1) {
+      // We found it.
+      break;
+    }
+
     tries++
 
     if (DEBUG) {
-      logger(fmt.Sprintf("Is pair alive? pair_name: %v (%v retries)\n",my_meta["pair_name"],tries))
+      logger(fmt.Sprintf("Is pair alive? pair_name: %v (%v retries)\n",linki.Pair_name,tries))
     }
 
     // We either timeout, or, we're alive.
-    if (tries >= ALIVE_WAIT_RETRIES || i_am_alive) {
+    if (tries >= ALIVE_WAIT_RETRIES) {
       return fmt.Errorf("Timeout: could not find that pair container is alive via metadata in %v tries", tries)
     }
 
     // Wait for however long.
     time.Sleep(ALIVE_WAIT_SECONDS * time.Second)
 
-  }
-
-  // Alright, given that... we should have a valid pair.
-  // So let's pick up the pair container id.
-  pairid_err, pair_containerid := getContainerIDByName(my_meta["pair_name"])
-  if (pairid_err != nil) {
-    return pairid_err
   }
 
   // Now, we can probably rock out all the 
@@ -287,12 +311,20 @@ func ratchet(argif string,containerid string, linki LinkInfo) error {
   // !trace !bang
   // This is how you call up koko.
   // koko.VethCreator("foo","192.168.2.100/24","in1","bar","192.168.2.101/24","in2")
-  koko_err := koko.VethCreator(containerid,my_meta["local_ip"],my_meta["local_ifname"],pair_containerid,my_meta["pair_ip"],my_meta["pair_ifname"])
+  koko_err := koko.VethCreator(
+    containerid,
+    linki.Local_ip + "/24",
+    linki.Local_ifname,
+    pair_containerid,
+    linki.Pair_ip + "/24",
+    linki.Pair_ifname,
+  )
+
   if (koko_err != nil) {
     return koko_err
   }
 
-  return result
+  return nil
 
 }
 
@@ -329,20 +361,42 @@ func main() {
 
   initEtcd(os.Args[3])
 
-  for idx,element := range os.Args {
-    logger(fmt.Sprintf("arg[%v]: %v",idx,element))
-  }
-
-
-  os.Exit(1)
-
   if (DEBUG) {
     logger("[LOGGING ENABLED]")
     logger(fmt.Sprintf("Interface: %v",os.Args[1]))
     logger(fmt.Sprintf("ContainerID: %v",os.Args[2]))
+    // Inspect all the arguments.
+    for idx,element := range os.Args {
+      logger(fmt.Sprintf("arg[%v]: %v",idx,element))
+    }
   }
 
+  // Apr 29 00:43:30 cni unknown[14497]: ratchet-child: arg[0]: /home/centos/cni/bin/ratchet-child
+  // Apr 29 00:43:30 cni unknown[14499]: ratchet-child: arg[1]: eth0
+  // Apr 29 00:43:30 cni unknown[14503]: ratchet-child: arg[2]: a3a3cb196700434c0ff58f0c07c8c51b95d2404826a017799b49c6a61224af1a
+  // Apr 29 00:43:30 cni cat[14507]: ratchet-child: arg[3]: localhost
+  // Apr 29 00:43:30 cni cat[14510]: ratchet-child: arg[4]: primary-pod
+  // Apr 29 00:43:30 cni unknown[14515]: ratchet-child: arg[5]: primary-pod
+  // Apr 29 00:43:30 cni unknown[14518]: ratchet-child: arg[6]: primary-pod
+  // Apr 29 00:43:30 cni unknown[14522]: ratchet-child: arg[7]: 1.1.1.1
+  // Apr 29 00:43:30 cni cat[14524]: ratchet-child: arg[8]: 192.168.2.100
+  // Apr 29 00:43:30 cni cat[14527]: ratchet-child: arg[9]: in1
+  // Apr 29 00:43:30 cni cat[14532]: ratchet-child: arg[10]: pair-pod
+  // Apr 29 00:43:30 cni cat[14539]: ratchet-child: arg[13]: true
+  // Apr 29 00:43:30 cni unknown[14537]: ratchet-child: arg[12]: in2
+  // Apr 29 00:43:30 cni unknown[14534]: ratchet-child: arg[11]: 192.168.2.101
+
   linki := LinkInfo{}
+  linki.Pod_name = os.Args[4]
+  linki.Target_pod = os.Args[5]
+  linki.Target_container = os.Args[6]
+  linki.Public_ip = os.Args[7]
+  linki.Local_ip = os.Args[8]
+  linki.Local_ifname = os.Args[9]
+  linki.Pair_name = os.Args[10]
+  linki.Pair_ip = os.Args[11]
+  linki.Pair_ifname = os.Args[12]
+  linki.Primary = os.Args[13]
 
   err := ratchet(os.Args[1],os.Args[2],linki)
   if (err != nil) {
